# src/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Pick_IUser.id-or-email-or-name-or-created_at-or-updated_at_": {
				"properties": {
					"id": {
						"type": "integer",
						"format": "int32"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"email",
					"name",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"IGetUserRes": {
				"$ref": "#/components/schemas/Pick_IUser.id-or-email-or-name-or-created_at-or-updated_at_"
			}
		},
		"securitySchemes": {
			"basic": {
				"type": "apiKey",
				"name": "authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "gmail-extension-server",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Jimmy Sorza ",
			"email": "sorza.jimmy05@gmail.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/auth/oauth": {
			"get": {
				"operationId": "GetAuthURI",
				"responses": {
					"302": {
						"description": "Redirect"
					}
				},
				"description": "Go to OAuth Page",
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "state",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/callback": {
			"get": {
				"operationId": "AuthCallback",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "code",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "state",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/info": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IGetUserRes"
								}
							}
						}
					}
				},
				"description": "Returns user info",
				"security": [
					{
						"basic": []
					}
				],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}